refactor: changed command name and namespace
