+++ README.md
@@ -35,14 +35,13 @@ Please use [Swagger editor](https://editor.swagger.io/) to visualize it.
 
 The application currently offers the following CLI commands:
 
-| Command                               | Description                   | Links                                                    |
-| --------------------------------------|:------------------------------|:---------------------------------------------------------|
-| `roster:activate:line-item`           | Activate/Deactivate line items| [link](docs/cli/line-item-activator-command.md)             |
-| `roster:cache-warmup:lti-instance`    | LTI instance cache warming    | [link](docs/cli/lti-instance-cache-warmer-command.md)    |
-| `roster:cache-warmup:line-item`       | Line Item cache warming       | [link](docs/cli/line-item-cache-warmer-command.md)       |
-| `roster:cache-warmup:user`            | User cache warming            | [link](docs/cli/user-cache-warmer-command.md)            |
-| `roster:garbage-collector:assignment` | Assignment garbage collection | [link](docs/cli/assignment-garbage-collector-command.md) |
-| `roster:ingest:lti-instance`          | LTI instance ingestion        | [link](docs/cli/lti-instance-ingester-command.md)        |
-| `roster:ingest:line-item`             | Line item instance ingestion  | [link](docs/cli/line-item-ingester-command.md)           |
-| `roster:ingest:user`                  | User ingestion                | [link](docs/cli/user-ingester-command.md)                |
-| `roster:ingest:assignment`            | Assignment ingestion          | [link](docs/cli/assignment-ingester-command.md)          |
+| Command                                                 | Description                   | Links                                                           |
+| --------------------------------------------------------|:------------------------------|:----------------------------------------------------------------|
+| `roster:modify-entity:line-item:change-state`           | Activate/Deactivate line items| [link](docs/cli/modify-entity-line-item-change-state-command.md)|
+| `roster:cache-warmup:lti-instance`                      | LTI instance cache warming    | [link](docs/cli/lti-instance-cache-warmer-command.md)           |
+| `roster:cache-warmup:user`                              | User cache warming            | [link](docs/cli/user-cache-warmer-command.md)                   |
+| `roster:garbage-collector:assignment`                   | Assignment garbage collection | [link](docs/cli/assignment-garbage-collector-command.md)        |
+| `roster:ingest:lti-instance`                            | LTI instance ingestion        | [link](docs/cli/lti-instance-ingester-command.md)               |
+| `roster:ingest:line-item`                               | Line item instance ingestion  | [link](docs/cli/line-item-ingester-command.md)                  |
+| `roster:ingest:user`                                    | User ingestion                | [link](docs/cli/user-ingester-command.md)                       |
+| `roster:ingest:assignment`                              | Assignment ingestion          | [link](docs/cli/assignment-ingester-command.md)                 |
